+++ src/features/index.ts
@@ -1,1 +1,3 @@
 export * from "./dreamsAdEngine/";
+export * from "./targeting/";
+export * from "./config/";