+++ types/features/config/index.d.ts
@@ -1,0 +1,3 @@
+export { DreamsAdConfig } from "./config.provider";
+export { DEFAULT_SLOTS } from "./default-mappings";
+export type { AdConfigInit, LazyLoadConfig, SlotConfig, } from "./config.types";