+++ vite.config.ts
@@ -1,22 +1,21 @@
 import { defineConfig, loadEnv } from "vite";
 import { resolve } from "node:path";
 
-// https://vitejs.dev/config/
 export default ({ mode }) => {
 	process.env = { ...process.env, ...loadEnv(mode, process.cwd()) };
 
 	return defineConfig({
-		base: process.env.VITE_NORA_ENV === "production" ? "./" : "./",
 		build: {
+			lib: {
+				entry: resolve(__dirname, "src/main.ts"),
+				name: "DreamsAdEngine",
+				fileName: "dreams-ad-engine",
+				formats: ["es"],
+			},
 			rollupOptions: {
-				input: {
-					main: resolve(__dirname, "index.html"),
-				},
+				external: [],
 				output: {
-					manualChunks: undefined,
 					assetFileNames: "dreams-ad-engine.[ext]",
-					entryFileNames: "dreams-ad-engine.js",
-					chunkFileNames: "dreams-ad-engine.js",
 				},
 			},
 			cssCodeSplit: false,